+++ the_flip/apps/parts/apps.py
@@ -9,7 +9,3 @@ class PartsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "the_flip.apps.parts"
     verbose_name = "Parts Management"
-
-    def ready(self):
-        """Import signals when the app is ready."""
-        from the_flip.apps.parts import signals  # noqa: F401
